# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIError:
    properties:
      details: {}
      message:
        type: string
    type: object
...
  handler.TipResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      relevance:
        type: integer
      source:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: contato@example.com
    name: Equipe Golang Finance
  description: API para gestão financeira pessoal com suporte a OCR de recibos.
  title: Golang Finance API
  version: "1.0"
paths:
...
  /tips:
    get:
      description: Retorna dicas ordenadas por relevância para o usuário autenticado
      parameters:
      - description: Mês (1-12)
        in: query
        name: month
        type: integer
      - description: Ano
        in: query
        name: year
        type: integer
      - description: true para recalcular dicas antes de responder
        in: query
        name: refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TipResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Listar dicas financeiras
      tags:
      - Dicas
  /tips/generate:
    post:
      description: Recalcula dicas com base nos gastos recentes e Gemini. Usa heurísticas
        caso o modelo não esteja disponível.
      parameters:
      - description: Mês (1-12)
        in: query
        name: month
        type: integer
      - description: Ano
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TipResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Gerar novas dicas financeiras
      tags:
      - Dicas
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"